producer/requirements.txt
requests
python-dotenv
kafka-python
